feat: full address system, Nominatim map search, toast notifications, cleanup
Backend:
- Replace lokasi_pkm table with address fields on pengajuan (provinsi, kota_kabupaten, kecamatan, kelurahan_desa, alamat_lengkap, latitude, longitude)
- Add soft deletes to users and aktivitas
- Add rate limiting on auth routes (10/min per IP)
- Fix UserController role validation (admin,dosen,masyarakat)
- Fix ArsipController validation to match seeder values
- Fix Pengajuan::testimoni() hasManyThrough relationship
- Fix missing imports in bootstrap/app.php
- Clean up PengajuanUserController (remove orphaned methods)
- Clean up TestimoniController (remove orphaned show method)
- Add global toast notifications via Inertia flash messages

Frontend:
- Convert all .jsx to .tsx
- Remove hardcoded PKM data (fetch from DB)
- Add Nominatim search to admin map picker (Aktivitas Detail)
- Replace file uploads with URL text fields for proposal/surat_permohonan
- Replace all alert()/confirm() with styled ConfirmDialog + inline validation
- Add global flash-to-toast in AdminLayout + DefaultLayout
- Delete orphaned files (UserLayout, MapDashboard, duplicate Dashboard)
- Remove Leaflet attribution labels
- Update user form with full address fields instead of lokasi dropdown

+++ database/migrations/2026_03_23_000008_create_testimoni_table.php
@@ -4,7 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration 
+return new class extends Migration
 {
     public function up(): void
     {
@@ -12,7 +12,7 @@ return new class extends Migration
             $table->id('id_testimoni');
             $table->foreignId('id_aktivitas')->constrained('aktivitas', 'id_aktivitas')->onDelete('cascade');
             $table->string('nama_pemberi');
-            $table->integer('rating')->default(5);
+            $table->unsignedSmallInteger('rating')->default(5);
             $table->text('pesan_ulasan')->nullable();
             $table->timestamps();
         });
